Create aircraft inputs example and update SimConnect
Added a new file `examples/aircraft_inputs/main.rs` that provides an example of handling aircraft inputs such as "GEAR_UP", "GEAR_DOWN", and "PARKING_BRAKES". The example includes the setup of a hashmap for input IDs to events cross-referencing and user console input for event triggering. Moreover, SimConnect's version was updated to 0.3.1.